1. Added support for JSON format log and daily performance output. 2. Fixed ordered type to match object_id type of UInt64 3. Added Performance tracer and modified performance calculations to use previous values

+++ Execution/Order.jl
@@ -13,7 +13,7 @@ include("../Security/Security.jl")
 Encapsulate the order characteristics    
 """
 type Order 
-  id::Integer # unique order id for submission/tracking
+  id::UInt64 # unique order id for submission/tracking
   securitysymbol::SecuritySymbol # security symbol (NIFTY, VOLTAS)
   quantity::Int64 # value of position change targeted
   remainingquantity::Int64
